uctMaths/requirements.txt
diff-match-patch
Django==4.2.7
django-allauth
django-import-export
docutils
#html5lib==1.1
html5lib
oauthlib
Pillow
xhtml2pdf
psycopg2
psycopg2-binary
PyPDF2
python-openid
pytz
reportlab
requests
requests-oauthlib
six
tablib
xlwt
